Update part4a.md
Closed connections in the "notes.js" router file using "mongoose.connection.close()". Online documentation states - "While it is not necessary to explicitly close the connection in mongoose, it is considered best practice". In the section "Part 3C - Creating and saving objects", it is recommended that we close the connection pool manually.